move: aiService에 있던 ai-resume 테스트를 test 폴더로 이동
